refactor(terminal): 拆出历史指令弹窗 hook 与组件
- 新增 components/terminal/：useTerminalHistory（历史列表加载/清空/删除、搜索过滤、键盘导航、弹窗定位与外点关闭）、TerminalHistoryPopup（历史指令弹窗）
- Terminal.tsx 1311 -> 999 行
